requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
mediapipe>=0.10.0
onnxruntime>=1.16.0
faiss-cpu>=1.7.4
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
supabase>=2.0.0
python-dotenv>=1.0.0
prometheus-client>=0.20.0
